P1-7: Decimal money and unit-safe quantities end to end
Every persisted money/quantity column (Trade.volume/fixed_price/strike_price
/premium, Position.net_volume/avg_fixed_price, ValuationResult.mtm_value
/realized_pnl/unrealized_pnl, VarResult.var_value, SensitivityResult.delta_value,
BookLimit.threshold, LimitBreach.threshold/observed_value, MarketDataPoint.price,
CurvePoint.price) was already Numeric(p,s) at the DB level -- SQLAlchemy has always
returned a real Decimal for those. What wasn't true: the Python-side Mapped[float]
annotations and the service-layer arithmetic on those values lied about that,
casting to float at dozens of call sites the instant a value came off the ORM. That
meant every sum/weighted-average over more than a handful of trades --
ValuationService.build_positions' net-volume/avg-price rollup chief among them,
since it feeds MTM, VaR, the delta ladder, stress tests, and pre-trade volume-limit
enforcement simultaneously -- accumulated ordinary binary-float rounding error.

app.common.money is the fix: Mapped[Decimal] on every column above (quant model
inputs/outputs with no persisted, summed-across-trades exactness requirement --
option_volatility, bump_size, pnl_impact -- deliberately stay float);
build_positions/attribute_pnl/limit-threshold comparisons now do their
summation/comparison in Decimal throughout; to_decimal() is the one clean
conversion point for a quant module's float output (Black-76, VaR, sensitivities,
curve bootstrap) crossing back into exact arithmetic at persistence; the audit log
stringifies Decimal values instead of float-casting them (json.dumps can't encode
Decimal, and an append-only compliance record needs the exact value); and the API
layer's MoneyDecimal type keeps Decimal exactness on the Python side while still
serializing as a plain JSON number on the wire (via a PlainSerializer, not Pydantic
v2's own default of a JSON string) -- so the actual bug (repeated arithmetic losing
precision) is fixed without an unrelated frontend contract change.

New tests: app.common.money's to_decimal/format_decimal; a direct proof that these
specific inputs drift under naive float summation and don't under
build_positions' Decimal arithmetic; Decimal-typed pnl_attribution unit tests.
Existing golden/unit tests updated to construct Decimal-typed fixtures where they
previously used plain float/int literals. Validated against real Postgres (a
Decimal column round-trips exactly through asyncpg, not just SQLite) and the real
Postgres+Redis worker/limit-concurrency integration tests.

Docs: ARCHITECTURE.md gains a "Decimal money and unit-safe quantities" section.

+++ backend/app/modules/limits/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import Date, DateTime, ForeignKey, Numeric, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column
@@ -21,7 +22,11 @@ class BookLimit(Base):
     book_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("books.id"), nullable=False)
     commodity: Mapped[Commodity] = mapped_column(String(30), nullable=False)
     limit_type: Mapped[LimitType] = mapped_column(String(10), nullable=False)
-    threshold: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
+    # Decimal -- compared directly against exact Decimal volumes/VaR values at
+    # enforcement time (see LimitService); a float threshold compared against a float
+    # observed value is exactly the kind of boundary comparison that can misfire by a
+    # rounding hair right at the limit.
+    threshold: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False)
     # Only meaningful for VAR limits; ignored for VOLUME.
     confidence_level: Mapped[int] = mapped_column(nullable=False, default=95)
 
@@ -47,8 +52,8 @@ class LimitBreach(Base):
     book_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("books.id"), nullable=False)
     commodity: Mapped[Commodity] = mapped_column(String(30), nullable=False)
     limit_type: Mapped[LimitType] = mapped_column(String(10), nullable=False)
-    threshold: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
-    observed_value: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
+    threshold: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False)
+    observed_value: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False)
     as_of_date: Mapped[date] = mapped_column(Date, nullable=False)
     status: Mapped[LimitBreachStatus] = mapped_column(
         String(20), nullable=False, default=LimitBreachStatus.OPEN
