+++ backend/app/modules/valuation/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import (
     Date,
@@ -76,8 +77,8 @@ class Position(Base):
     book_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("books.id"), nullable=False)
     commodity: Mapped[Commodity] = mapped_column(String(30), nullable=False)
     delivery_month: Mapped[date] = mapped_column(Date, nullable=False)
-    net_volume: Mapped[float] = mapped_column(Numeric(18, 4), nullable=False)
-    avg_fixed_price: Mapped[float] = mapped_column(Numeric(18, 6), nullable=False)
+    net_volume: Mapped[Decimal] = mapped_column(Numeric(18, 4), nullable=False)
+    avg_fixed_price: Mapped[Decimal] = mapped_column(Numeric(18, 6), nullable=False)
     as_of_date: Mapped[date] = mapped_column(Date, nullable=False)
 
 
@@ -121,9 +122,9 @@ class ValuationResult(Base):
     delivery_month: Mapped[date | None] = mapped_column(Date, nullable=True)
     as_of_date: Mapped[date] = mapped_column(Date, nullable=False)
     curve_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("forward_curves.id"), nullable=False)
-    mtm_value: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
-    realized_pnl: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False, default=0)
-    unrealized_pnl: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
+    mtm_value: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False)
+    realized_pnl: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False, default=0)
+    unrealized_pnl: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False)
     currency: Mapped[Currency] = mapped_column(String(5), nullable=False, default=Currency.USD)
     computed_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
